Support passing NULL Fix objects in, added CheckName to Set-IssueFix

+++ PoshIssues.psm1
@@ -144,49 +144,52 @@ function Write-IssueFix {
                 [Switch] $NoClobber
 	)
 	Process {
-                #Create an object to save as JSON
-                $_fix = @{
-                        "id" = $Fix.id;
-                        "sequenceNumber" = $Fix.sequenceNumber;
-                        "checkName" = $Fix.checkName;
-                        "fixCommandBase64" = [Convert]::ToBase64String([Text.Encoding]::Unicode.GetBytes($Fix.fixCommand));
-                        "fixDescription" = $Fix.fixDescription;
-                        "fixResults" = $Fix.fixResults;
-                        "statusInt" = $Fix._status;
-                        "notificationCount" = $Fix.notificationCount;
-                        "creationDateTime" = $Fix.creationDateTime;
-                        "statusDateTime" = $Fix.statusDateTime
-                }
-                $_path = ""
-                If ($DatabasePath) {
-                        #Save to database path, overwriting only if Force
-                        $_fix.Add("databasePath", $DatabasePath)
-                        Add-Member -InputObject $Fix -MemberType NoteProperty -Name "databasePath" -Value $DatabasePath -Force
-                        
-                        if (!(Test-Path $DatabasePath)) {
-                                New-Item $DatabasePath -ItemType Directory
+                #Make sure we got a fix passed
+                if ($Fix) {
+                        #Create an object to save as JSON
+                        $_fix = @{
+                                "id" = $Fix.id;
+                                "sequenceNumber" = $Fix.sequenceNumber;
+                                "checkName" = $Fix.checkName;
+                                "fixCommandBase64" = [Convert]::ToBase64String([Text.Encoding]::Unicode.GetBytes($Fix.fixCommand));
+                                "fixDescription" = $Fix.fixDescription;
+                                "fixResults" = $Fix.fixResults;
+                                "statusInt" = $Fix._status;
+                                "notificationCount" = $Fix.notificationCount;
+                                "creationDateTime" = $Fix.creationDateTime;
+                                "statusDateTime" = $Fix.statusDateTime
                         }
-                        if (!(Test-Path "$($DatabasePath)\Fixes")) {
-                                New-Item "$($DatabasePath)\Fixes" -ItemType Directory
+                        $_path = ""
+                        If ($DatabasePath) {
+                                #Save to database path, overwriting only if Force
+                                $_fix.Add("databasePath", $DatabasePath)
+                                Add-Member -InputObject $Fix -MemberType NoteProperty -Name "databasePath" -Value $DatabasePath -Force
+                                
+                                if (!(Test-Path $DatabasePath)) {
+                                        New-Item $DatabasePath -ItemType Directory
+                                }
+                                if (!(Test-Path "$($DatabasePath)\Fixes")) {
+                                        New-Item "$($DatabasePath)\Fixes" -ItemType Directory
+                                }
+                                $_path = "$($DatabasePath)\Fixes\$($Fix.id).json"
+                        } else {
+                                #Save to path, overwriting only if Force
+                                $_fix.Add("path", $Path)
+                                Add-Member -InputObject $Fix -MemberType NoteProperty -Name "path" -Value $Path -Force
+                                $_path = $Path
                         }
-                        $_path = "$($DatabasePath)\Fixes\$($Fix.id).json"
-                 } else {
-                        #Save to path, overwriting only if Force
-                        $_fix.Add("path", $Path)
-                        Add-Member -InputObject $Fix -MemberType NoteProperty -Name "path" -Value $Path -Force
-                        $_path = $Path
-                }
 
-                #If the file exists AND NoClobber is true not write the file
-                if ((Test-Path $_path) -and ($NoClobber)) {
-                        Write-Verbose "JSON file already exists at '$_path' and NoClobber is set."
-                } else {
-                        $_json = ConvertTo-Json -InputObject $_fix
-                        Out-File -FilePath $_path -Force:$true -InputObject $_json
-                        Write-Verbose "JSON saved to '$_path'."
-                }
+                        #If the file exists AND NoClobber is true not write the file
+                        if ((Test-Path $_path) -and ($NoClobber)) {
+                                Write-Verbose "JSON file already exists at '$_path' and NoClobber is set."
+                        } else {
+                                $_json = ConvertTo-Json -InputObject $_fix
+                                Out-File -FilePath $_path -Force:$true -InputObject $_json
+                                Write-Verbose "JSON saved to '$_path'."
+                        }
 
-                Write-Output $Fix
+                        Write-Output $Fix
+                }
 	}
 }
 
@@ -224,28 +227,31 @@ function Remove-IssueFix {
                 [String] $Path
 	)
 	Process {
-                $_path = ""
-                #Calculate path
-                If ($DatabasePath) {
-                        $_path = "$($DatabasePath)\Fixes\$($Fix.id).json"
-                 } else {
-                        $_path = $Path
-                }
-                if ($_path -eq "") {
-                        Write-Error "Unable to determine path to saved Fix"
-                } else {
-                        if (Test-Path $_path) {
-                                if ($PSCmdlet.ShouldProcess("Remove $($Fix.fixDescription) from file/database?")) {
-                                        #Delete the JSON file
-                                        Write-Verbose "Removed $_path"
-                                        Remove-Item $_path
-                                }
+                #Make sure we got a fix passed
+                if ($Fix) {
+                        $_path = ""
+                        #Calculate path
+                        If ($DatabasePath) {
+                                $_path = "$($DatabasePath)\Fixes\$($Fix.id).json"
                         } else {
-                                Write-Warning "Saved Fix JSON file not found at $_path"
+                                $_path = $Path
                         }
-                }
+                        if ($_path -eq "") {
+                                Write-Error "Unable to determine path to saved Fix"
+                        } else {
+                                if (Test-Path $_path) {
+                                        if ($PSCmdlet.ShouldProcess("Remove $($Fix.fixDescription) from file/database?")) {
+                                                #Delete the JSON file
+                                                Write-Verbose "Removed $_path"
+                                                Remove-Item $_path
+                                        }
+                                } else {
+                                        Write-Warning "Saved Fix JSON file not found at $_path"
+                                }
+                        }
 
-                Write-Output $Fix
+                        Write-Output $Fix
+                }
 	}
 }
 
@@ -293,40 +299,43 @@ function Archive-IssueFix {
                 [Switch] $Force
 	)
 	Process {
-                $_path = ""
-                $_destinationPath = ""
-                #Calculate path
-                If ($DatabasePath) {
-                        $_path = "$($DatabasePath)\Fixes\$($Fix.id).json"
-                        $_destinationPath = "$($DatabasePath)\Fixes\Archive\$($Fix.id)_$(Get-Date -Format yyyyMMddHHmmss).json"
-                        if (!(Test-Path $DatabasePath)) {
-                                New-Item $DatabasePath -ItemType Directory
-                        }
-                        if (!(Test-Path "$($DatabasePath)\Fixes")) {
-                                New-Item "$($DatabasePath)\Fixes" -ItemType Directory
-                        }
-                        if (!(Test-Path "$($DatabasePath)\Fixes\Archive")) {
-                                New-Item "$($DatabasePath)\Fixes\Archive" -ItemType Directory
-                        }
-                 } else {
-                        $_path = $Path
-                        $_destinationPath = $ArchivePath
-                }
-                if ($_path -eq "") {
-                        Write-Error "Unable to determine path to saved Fix"
-                } else {
-                        if (Test-Path $_path) {
-                                if ($PSCmdlet.ShouldProcess("Move $($Fix.fixDescription) to $_destinationPath?")) {
-                                        #Move the JSON file
-                                        Write-Verbose "Moved $_path to $_destinationPath"
-                                        Move-Item -Path $_path -Destination $_destinationPath -Force:$Force
+                #Make sure we got a fix passed
+                if ($Fix) {
+                        $_path = ""
+                        $_destinationPath = ""
+                        #Calculate path
+                        If ($DatabasePath) {
+                                $_path = "$($DatabasePath)\Fixes\$($Fix.id).json"
+                                $_destinationPath = "$($DatabasePath)\Fixes\Archive\$($Fix.id)_$(Get-Date -Format yyyyMMddHHmmss).json"
+                                if (!(Test-Path $DatabasePath)) {
+                                        New-Item $DatabasePath -ItemType Directory
+                                }
+                                if (!(Test-Path "$($DatabasePath)\Fixes")) {
+                                        New-Item "$($DatabasePath)\Fixes" -ItemType Directory
+                                }
+                                if (!(Test-Path "$($DatabasePath)\Fixes\Archive")) {
+                                        New-Item "$($DatabasePath)\Fixes\Archive" -ItemType Directory
                                 }
                         } else {
-                                Write-Warning "Saved Fix JSON file not found at $_path"
+                                $_path = $Path
+                                $_destinationPath = $ArchivePath
                         }
-                }
+                        if ($_path -eq "") {
+                                Write-Error "Unable to determine path to saved Fix"
+                        } else {
+                                if (Test-Path $_path) {
+                                        if ($PSCmdlet.ShouldProcess("Move $($Fix.fixDescription) to $_destinationPath?")) {
+                                                #Move the JSON file
+                                                Write-Verbose "Moved $_path to $_destinationPath"
+                                                Move-Item -Path $_path -Destination $_destinationPath -Force:$Force
+                                        }
+                                } else {
+                                        Write-Warning "Saved Fix JSON file not found at $_path"
+                                }
+                        }
 
-                Write-Output $Fix
+                        Write-Output $Fix
+                }
 	}
 }
 
@@ -402,36 +411,44 @@ function Set-IssueFix {
                 [Parameter(Mandatory=$false,Position=1,ValueFromPipeline=$false,ValueFromPipelineByPropertyName=$true)]
                 [String] $FixDescription,
                 [Parameter(Mandatory=$false,Position=2,ValueFromPipeline=$false,ValueFromPipelineByPropertyName=$true)]
-                [IssueFixStatus] $Status,
+                [String] $CheckName = "",
                 [Parameter(Mandatory=$false,Position=3,ValueFromPipeline=$false,ValueFromPipelineByPropertyName=$true)]
-                [System.Int64] $NotificationCount,
+                [IssueFixStatus] $Status,
                 [Parameter(Mandatory=$false,Position=4,ValueFromPipeline=$false,ValueFromPipelineByPropertyName=$true)]
+                [System.Int64] $NotificationCount,
+                [Parameter(Mandatory=$false,Position=5,ValueFromPipeline=$false,ValueFromPipelineByPropertyName=$true)]
                 [System.Int64] $SequenceNumber
 	)
 	Begin {
         #Put begining stuff here
 	}
 	Process {
-                if ($PSCmdlet.ShouldProcess("Change $($Fix.fixDescription)?")) {
-                        if ($FixDescription) {
-                                $Fix.fixDescription = $FixDescription
-                        }
-                        if ($Status) {
-                                if (($Status -ge 0) -and ($Status -le 4)) {
-                                        $Fix._status = $Status
-                                        $Fix.statusDateTime = Get-Date
-                                } else {
-                                        Write-Warning "Invalid status value"
+                #Make sure we got a fix passed
+                if ($Fix) {
+                        if ($PSCmdlet.ShouldProcess("Change $($Fix.fixDescription)?")) {
+                                if ($CheckName) {
+                                        $Fix.CheckName = $CheckName
+                                }
+                                if ($FixDescription) {
+                                        $Fix.fixDescription = $FixDescription
+                                }
+                                if ($Status) {
+                                        if (($Status -ge 0) -and ($Status -le 4)) {
+                                                $Fix._status = $Status
+                                                $Fix.statusDateTime = Get-Date
+                                        } else {
+                                                Write-Warning "Invalid status value"
+                                        }
+                                }
+                                If ($NotificationCount) {
+                                        $Fix.notificationCount = $NotificationCount
+                                }
+                                if ($SequenceNumber) {
+                                        $Fix.sequenceNumber = $SequenceNumber
                                 }
                         }
-                        If ($NotificationCount) {
-                                $Fix.notificationCount = $NotificationCount
-                        }
-                        if ($SequenceNumber) {
-                                $Fix.sequenceNumber = $SequenceNumber
-                        }
+                        Write-Output $Fix
                 }
-                Write-Output $Fix
 	}
 	End {
         #Put end here
@@ -445,11 +462,14 @@ function Approve-IssueFix {
 		[PSObject] $Fix
 	)
 	Process {
-                if ($PSCmdlet.ShouldProcess("Change $($Fix.fixDescription) from $($Fix.status) to Complete?")) {
-                        $Fix._status = 0
-                        $Fix.statusDateTime = Get-Date
+                #Make sure we got a fix passed
+                if ($Fix) {
+                        if ($PSCmdlet.ShouldProcess("Change $($Fix.fixDescription) from $($Fix.status) to Complete?")) {
+                                $Fix._status = 0
+                                $Fix.statusDateTime = Get-Date
+                        }
+                        Write-Output $Fix
                 }
-                Write-Output $Fix
 	}
 }
 
@@ -460,11 +480,14 @@ function Deny-IssueFix {
 		[PSObject] $Fix
 	)
 	Process {
-                if ($PSCmdlet.ShouldProcess("Change $($Fix.fixDescription) from $($Fix.status) to Complete?")) {
-                        $Fix._status = 4
-                        $Fix.statusDateTime = Get-Date
+                #Make sure we got a fix passed
+                if ($Fix) {
+                        if ($PSCmdlet.ShouldProcess("Change $($Fix.fixDescription) from $($Fix.status) to Complete?")) {
+                                $Fix._status = 4
+                                $Fix.statusDateTime = Get-Date
+                        }
+                        Write-Output $Fix
                 }
-                Write-Output $Fix
 	}
 }
 
@@ -479,24 +502,27 @@ function Invoke-IssueFix {
                 [Switch] $NoNewScope
 	)
 	Process {
-                if (($Fix.status -eq 0) -or $Force) {
-                        if ($PSCmdlet.ShouldProcess("Invoke $($Fix.fixDescription) from $($Fix.checkName) by running $($Fix.fixCommand)?")) {
-                                Add-Member -InputObject $Fix -MemberType NoteProperty -Name "fixResults" -Value "" -Force
-                                try {
-                                        $Fix.fixResults = [String] (Invoke-Command -ScriptBlock $fix.fixCommand -NoNewScope:$NoNewScope)
-                                        $Fix.status = 2 #Complete
-                                        Write-Verbose "$($Fix.checkName): $($Fix.fixDescription) complete with following results: $($Fix.fixResults)"
-                                } catch {
-                                        #Error
-                                        $Fix.fixResults = [String] $_.Exception.Message
-                                        $Fix.status = 3 #Error
-                                        Write-Verbose "$($Fix.checkName): $($Fix.fixDescription) errored with following error: $($Fix.fixResults)"
-                                } finally {
-                                        $Fix.statusDateTime = Get-Date
+                #Make sure we got a fix passed
+                if ($Fix) {
+                        if (($Fix.status -eq 0) -or $Force) {
+                                if ($PSCmdlet.ShouldProcess("Invoke $($Fix.fixDescription) from $($Fix.checkName) by running $($Fix.fixCommand)?")) {
+                                        Add-Member -InputObject $Fix -MemberType NoteProperty -Name "fixResults" -Value "" -Force
+                                        try {
+                                                $Fix.fixResults = [String] (Invoke-Command -ScriptBlock $fix.fixCommand -NoNewScope:$NoNewScope)
+                                                $Fix.status = 2 #Complete
+                                                Write-Verbose "$($Fix.checkName): $($Fix.fixDescription) complete with following results: $($Fix.fixResults)"
+                                        } catch {
+                                                #Error
+                                                $Fix.fixResults = [String] $_.Exception.Message
+                                                $Fix.status = 3 #Error
+                                                Write-Verbose "$($Fix.checkName): $($Fix.fixDescription) errored with following error: $($Fix.fixResults)"
+                                        } finally {
+                                                $Fix.statusDateTime = Get-Date
+                                        }
                                 }
                         }
+                        Write-Output $Fix
                 }
-                Write-Output $Fix
 	}
 }
